[New] Embed vanilla datapack files into the executable
- `1.x.x-data-json.qrc` (1.x.x is a game version) files specify which
files are embedded.
- Not all vanilla files are included. If a file is the same as the
corresponding file of its previous version, it won't be included.
- `vanilla_lookup.json` (or `vanilla_lookup.cbor`) files are used for
finding out which version a file belongs to. If an internal relative
file path (alias path) doesn't exist in the lookup file and
QFile::exists() returns false, the file does not exist.
- The system is currently a prototype and is subject to change.

game.h:
- Add functions to check if a vanilla file exists and to get its real
path (if exists).

Highlighter:
- Now also try to locate namespaced IDs in the `advancements`,
`loot_tables`, and `recipes` folders of the vanilla datapack.

Signed-off-by: IoeCmcomc <53734763+IoeCmcomc@users.noreply.github.com>