[Refactor] Rename `resource` folder to `res`, `minecraft` to `mc`
The changes are done to make the program compilable using the MinGW toolchain. The reason is that some included file paths are too long (more than 260 characters), causing MinGW's `make` to treat these files as not existing, halting the compilation process. It doesn't happen with the MSVC toolchain. Renaming these two folders will reduce the path lengths to below the maximum limits.